Make crypto.c compile/link with OpenSSL 3
- Initialize libssl using non-deprecated APIs

OpenSSL 3 deprecated `SSL_library_init` and `SSL_load_error_strings` in
favor of `OPENSSL_init_ssl`. Use `OPENSSL_init_ssl` when dealing with
OpenSSL 1.1 and newer to unbreak the build with OpenSSL 3.

- Move MD5 APIs to EVP_MD APIs

OpenSSL 3 deprecated all of the `MD5_`* APIs. Move to the equivalent
`EVP_MD`* APIs so the code doesn't need to be pinned down to 1.1
compatible APIs and uplifted at a later date.

Co-authored-by: Pierre Pronchery <pierre@freebsdfoundation.org>
Co-authored-by: Ed Maste <emaste@FreeBSD.org>
Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>
Sponsored by:	The FreeBSD Foundation