Terminate each local message by an empty line
According to RFC 4155, each message in the mbox database must be
terminated by an empty line, containing a single end-of-line marker.
That doesn't match the current behavior, where dma writes this EOL
marker at the same time with the following separator line on each
subsequent local delivery, which effectively changes the previous
message, and may leave the new one in a non-conformant state.